LPD-58785 Move naming from Database Schema Definition to Database Migration Schema Exporter.
