+++ modules/apps/portal-language/portal-language-lang/src/main/resources/content/Language.properties
@@ -5564,9 +5564,9 @@ day-view=Day View
 day-view-header-date-format=Day View Header Date Format
 days=Days
 days-abbreviation=d
+db-migration-schema-export-configuration-description=Exports PostgreSQL database schema definition to tables.sql and indexes.sql files.
+db-migration-schema-export-configuration-name=Database Migration Schema Export
 db-partition-configuration-name=Database Partition
-db-schema-definition-exporter-configuration-description=Exports PostgreSQL database schema definition to tables.sql and indexes.sql files.
-db-schema-definition-exporter-configuration-name=Database Schema Definition Exporter
 dd-yyyy=DD, YYYY
 ddd-mmm-d=ddd, MMM D
 ddl-record-local-service=DDL Record Local Service
@@ -7651,7 +7651,7 @@ feature.flag.LPD-20379.title=Commerce Classic Site Initializer
 feature.flag.LPD-20556.description=Enhance the publication toolbar to provide users with additional context for editors, including current publication details, real time collaboration status, and change history.
 feature.flag.LPD-20556.title=Additional Context in Publications Toolbar
 feature.flag.LPD-23840.description=System administrators can export the database schema definition for migration to PostgreSQL.
-feature.flag.LPD-23840.title=Database Schema Definition Exporter System Settings
+feature.flag.LPD-23840.title=Database Migration Schema Export System Settings
 feature.flag.LPD-29516.description=Users can upload images to the blog repository even when the workflow is active.
 feature.flag.LPD-29516.title=Image Uploads to Blog Repository with Active Workflow
 feature.flag.LPD-30204.description=Allow different pages on the same site to have style books based on different themes.
