+++ modules/apps/portal/portal-db-migration-schema-exporter-test/bnd.bnd
@@ -1,0 +1,3 @@
+Bundle-Name: Liferay Portal DB Migration Schema Exporter Test
+Bundle-SymbolicName: com.liferay.portal.db.migration.schema.exporter.test
+Bundle-Version: 1.0.0
