+++ modules/apps/portal/portal-db-migration-schema-exporter-test/test.properties
