+++ modules/apps/portal/portal-db-migration-schema-exporter/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Portal DB Migration Schema Exporter
+Bundle-SymbolicName: com.liferay.portal.db.migration.schema.exporter
+Bundle-Version: 4.0.3
+-dsannotations-options: inherit
