+++ modules/apps/portal/portal-db-migration-schema-exporter/build.gradle
