+++ modules/apps/portal/portal-db-migration-schema-exporter/src/main/resources/META-INF/module-log4j.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0"?>
+
+<Configuration strict="true">
+	<Loggers>
+		<Logger level="INFO" name="com.liferay.portal.db.migration.schema.exporter.internal.exporter.DBMigrationSchemaExport" />
+	</Loggers>
+</Configuration>
