--- modules/apps/portal/portal-db-schema-definition-test/bnd.bnd
@@ -1,3 +1,0 @@
-Bundle-Name: Liferay Portal DB Schema Definition Test
-Bundle-SymbolicName: com.liferay.portal.db.schema.definition.test
-Bundle-Version: 1.0.0
