--- modules/apps/portal/portal-db-schema-definition/bnd.bnd
@@ -1,4 +1,0 @@
-Bundle-Name: Liferay Portal DB Schema Definition
-Bundle-SymbolicName: com.liferay.portal.db.schema.definition
-Bundle-Version: 4.0.3
--dsannotations-options: inherit
