+++ test.properties
@@ -3303,8 +3303,8 @@
 
     test.batch.class.names.includes[db-partition]=\
         **/portal/kernel/**/CompanyThreadLocalTest.java,\
+        **/portal-db-migration-schema-exporter-test/**/*Test.java,\
         **/portal-db-partition-test/**/*Test.java,\
-        **/portal-db-schema-definition-test/**/*Test.java,\
         **/portal-tools-db-partition*/**/*Test.java,\
         **/testIntegration/**/UpgradeProcessDBPartitionTest.java,\
         **/testIntegration/**/company/service/test/*Test.java,\
@@ -6818,7 +6818,7 @@
         **/portal/kernel/upgrade/util/*Test.java,\
         **/portal-dao-test/**/*Test.java,\
         **/portal-dao-test/**/*Test.java,\
-        **/portal-db-schema-definition-test/**/*Test.java,\
+        **/portal-db-migration-schema-exporter-test/**/*Test.java,\
         **/portal-upgrade-impl/**/*Test.java,\
         **/portal-upgrade-test/**/*Test.java,\
         **/portal-verify-test/**/Preupgrade*Test.java,\
